requirements.txt
mne>=1.6
brainflow>=5.12
scipy>=1.11
numpy>=1.24
pandas>=2.0
pyarrow>=14.0

xgboost>=2.0
scikit-learn>=1.3
shap>=0.44
joblib>=1.3

streamlit>=1.32
plotly>=5.18
fpdf2>=2.7
Pillow>=10.0

pyyaml>=6.0
python-dotenv>=1.0
loguru>=0.7